This patch should fix a few more of the pesky warnings that shows up on an x86_64 machine with gcc and when built with --enable-strict. Also, this patch should fix the -EACCES error (permission denied) that Ekow reported which was caused by broken permission checking during the recent acl merges.
SVN revision 4650

+++ src/kernel/linux-2.6/devpvfs2-req.c
@@ -387,7 +387,7 @@ static ssize_t pvfs2_devreq_writev(
                     {
                         pvfs2_print("WARNING: pvfs2_iocb from op"
                                 "has invalid fields! %p, %p(%p), %d\n",
-                                x->buffer, x->op, op, x->bytes_to_be_copied);
+                                x->buffer, x->op, op, (int) x->bytes_to_be_copied);
                     }
                     else
                     {
